requirements.txt
super-image
Pillow
pytest
